Remove redundant float casts in burst damage registration methods
Agent-Logs-Url: https://github.com/rpillai25/PitHero/sessions/21df99da-a9d9-4c22-866d-552b74a50876

+++ PitHero/ECS/Components/HeroComponent.cs
@@ -460,7 +460,7 @@ namespace PitHero.ECS.Components
             if (LinkedHero == null || LinkedHero.MaxHP <= 0)
                 return;
             float burstThreshold = LinkedHero.MaxHP * GameConfig.BurstDamageThresholdPercent;
-            if ((float)damage >= burstThreshold)
+            if (damage >= burstThreshold)
                 _heroBurstDamageTriggered = true;
         }
 
@@ -477,7 +477,7 @@ namespace PitHero.ECS.Components
             if (mercComp?.LinkedMercenary == null || mercComp.LinkedMercenary.MaxHP <= 0)
                 return;
             float burstThreshold = mercComp.LinkedMercenary.MaxHP * GameConfig.BurstDamageThresholdPercent;
-            if ((float)damage >= burstThreshold)
+            if (damage >= burstThreshold)
                 _burstDamageMercEntityIds.Add(mercEntity.Id);
         }
